pyproject.toml
[project]
name = "plc-toolkits-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bitstring>=4.3.0",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "python-snap7>=2.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.3",
    "schedule>=1.2.2",
]
